AI can now choose a target word even if they go first.
Added whitelist and blacklist user dictionaries. functionality not yet implemented.

build_dictionary function now maintains the word order of the frequency-ordered dictionary, but removes all non-standard words in the process.

Began implementing user options at the end of a game to allow replay and removal of words via addition to the blacklist.